backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF Processing
pypdf>=3.17.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Vector Database
chromadb>=0.4.22

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
